+++ packages/react-native-editor/CHANGELOG.md
@@ -11,7 +11,14 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+## 1.77.0
+
 -   [*] [a11y] Improve text read by screen readers for BottomSheetSelectControl [#41036]
+-   [*] Add 'Insert from URL' option to Image block [#40334]
+
+## 1.76.2
+
+-   [*] Ensure post title gets focused when is notified from native side [#41371]
 
 ## 1.76.1